Update enum and type storage
