+++ apps/frontend/src/components/form/MedicalForm.tsx
@@ -13,29 +13,11 @@ import {
 } from '@chakra-ui/react';
 import { PhoneIcon } from '@chakra-ui/icons';
 import { Field, FieldProps } from 'formik';
+import {
+  CancersDropdownValues,
+  HospitalsDropdownValues,
+} from '../../enums/DropdownValues';
 
-export enum CancersDropdownValues {
-  BRAINCANCERS = 'braincancers',
-  LEUKEMIAS = 'leukemias',
-  LYMPHOMAS = 'lymphomas',
-  THYROIDCANCER = 'thyroidCancer',
-  GERMCELLTUMORS = 'germCellGonadalTumors',
-  NEUROBLASTOMAS = 'neuroblastomas',
-  OSTEOSARCOMA = 'osteosarcoma',
-  OTHERCANCERS = 'otherCancers',
-}
-export enum HospitalsDropdownValues {
-  BOSHOSPITAL = 'bostonChildrenHospital',
-  CTHOPSITAL = 'connecticutChildrenHospital',
-  DANAHOSPITAL = 'danaFarberHospital',
-  HASBROHOSPITAL = 'hasbroChildrenHospital',
-  HOPEHOSPITAL = 'hopeHealthHospital',
-  MASSGENHOSPITAL = 'massGeneralHospital',
-  TUFTSHOSPITAL = 'tuftsMedicalHospital',
-  UMASSHOSPITAL = 'umassHospital',
-  VERMONTHOSPITAL = 'universityOfVermontHospital',
-  OTHERHOSPITALS = 'otherHospitals',
-}
 interface MedicalFormValues {
   childDiagnosis: string;
   otherCancer: string;
