+++ apps/frontend/src/enums/DropdownValues.ts
@@ -1,0 +1,22 @@
+export enum CancersDropdownValues {
+  BRAINCANCERS = 'braincancers',
+  LEUKEMIAS = 'leukemias',
+  LYMPHOMAS = 'lymphomas',
+  THYROIDCANCER = 'thyroidCancer',
+  GERMCELLTUMORS = 'germCellGonadalTumors',
+  NEUROBLASTOMAS = 'neuroblastomas',
+  OSTEOSARCOMA = 'osteosarcoma',
+  OTHERCANCERS = 'otherCancers',
+}
+export enum HospitalsDropdownValues {
+  BOSHOSPITAL = 'bostonChildrenHospital',
+  CTHOPSITAL = 'connecticutChildrenHospital',
+  DANAHOSPITAL = 'danaFarberHospital',
+  HASBROHOSPITAL = 'hasbroChildrenHospital',
+  HOPEHOSPITAL = 'hopeHealthHospital',
+  MASSGENHOSPITAL = 'massGeneralHospital',
+  TUFTSHOSPITAL = 'tuftsMedicalHospital',
+  UMASSHOSPITAL = 'umassHospital',
+  VERMONTHOSPITAL = 'universityOfVermontHospital',
+  OTHERHOSPITALS = 'otherHospitals',
+}
