+++ apps/frontend/src/enums/SortOrder.ts
@@ -1,0 +1,6 @@
+enum SortOrder {
+  'ASC',
+  'DESC',
+}
+
+export default SortOrder;
